fixed not working models tests

+++ migrations/sqls/20220305171135-mythical-worlds-table-up.sql
@@ -8,7 +8,7 @@ price INTEGER NOT NULL);
 
 CREATE TABLE Orders (id SERIAL PRIMARY  KEY,
 status VARCHAR(255) NOT NULL,
-user_id bigint references Users(id));
+user_id INTEGER references Users(id));
 
 CREATE TABLE Order_Products (id SERIAL PRIMARY KEY,quantity INTEGER ,order_id bigint references Orders(id),product_id bigint references Products(id));
